requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
cx-Oracle==8.0.1
Django==3.1
django-admin-interface==0.14.1
django-colorfield==0.3.2
django-crispy-forms==1.10.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-geoposition==0.3.0
django-pwa==1.0.10
django-rest-framework==0.1.0
djangorestframework==3.12.2
idna==2.10
numpy==1.16.4
Pillow==8.0.1
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4